imporved PCL support
