+++ src/SimpleAuth.Droid/SimpleAuth.Droid.csproj
@@ -55,6 +55,7 @@
     <Compile Include="ActivityStateRepository.cs" />
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="AuthStorage.cs" />
+    <Compile Include="TypeForwarded.cs" />
     <Compile Include="WebAuthenticatorActivity.cs" />
     <Compile Include="LoginActvity.cs" />
   </ItemGroup>
@@ -78,6 +79,5 @@
   -->
   <ItemGroup>
     <Folder Include="Properties\" />
-    <Folder Include="Resources\layout\" />
   </ItemGroup>
 </Project>
