+++ src/SimpleAuth.Droid/TypeForwarded.cs
@@ -1,0 +1,5 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: TypeForwardedTo(typeof(System.Console))]
+[assembly: TypeForwardedTo(typeof(System.Web.HttpUtility))]
+[assembly: TypeForwardedTo(typeof(System.Collections.Specialized.NameValueCollection))]
