+++ src/SimpleAuth.iOS/SimpleAuth.iOS.csproj
@@ -39,6 +39,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="BasicAuthController.cs" />
+    <Compile Include="TypeForwarded.cs" />
     <Compile Include="WebAuthenticatorViewController.cs" />
     <Compile Include="AuthStorage.cs" />
   </ItemGroup>
